replaced unsigned long long with size_t, long long with intmax_t and LLONG_MAX with SIZE_MAX >> 1

+++ lib/generics/src/hash_table.c
@@ -34,19 +34,15 @@ void table_destroy(struct hash_table *table) {
   if (!table) return;
   if (!table->entries) return;
 
-  for (unsigned long long i = 0; i < table->capacity; i++) {
+  for (size_t i = 0; i < table->capacity; i++) {
     // destroy all buckets in an entry
     struct entry *entry = vector_at(table->entries, i);
     for (struct node *bucket = entry->head; bucket; entry->head = bucket) {
       bucket = bucket->next;
 
-      if (table->destroy_key) {
-        table->destroy_key(entry->head->key);
-      }
+      if (table->destroy_key) { table->destroy_key(entry->head->key); }
 
-      if (table->destroy_value) {
-        table->destroy_value(entry->head->value);
-      }
+      if (table->destroy_value) { table->destroy_value(entry->head->value); }
 
       if (entry->head->key) free(entry->head->key);
       if (entry->head->value) free(entry->head->value);
@@ -62,22 +58,22 @@ bool table_empty(struct hash_table *table) {
   return table->num_of_elements == 0;
 }
 
-unsigned long long table_size(struct hash_table *table) {
+size_t table_size(struct hash_table *table) {
   if (!table) return 0;
   return table->num_of_elements;
 }
 
-unsigned long long table_capacity(struct hash_table *table) {
+size_t table_capacity(struct hash_table *table) {
   if (!table) return 0;
   return table->capacity;
 }
 
 /* used internally to hash the keys (slightly modified djd2 by Dan Bernstein)
  */
-static unsigned long long hash(const void *key, unsigned long long key_size) {
+static size_t hash(const void *key, size_t key_size) {
   const unsigned char *k = key;
-  unsigned long long hash = 5381;
-  for (unsigned long long i = 0; i < key_size; i++, k++) {
+  size_t hash = 5381;
+  for (size_t i = 0; i < key_size; i++, k++) {
     hash = hash * 33 + *k;
   }
   return hash;
@@ -89,9 +85,7 @@ static unsigned long long hash(const void *key, unsigned long long key_size) {
  * both node::key, node::value and node must be free'd. the function assumes
  * key
  * != NULL and key_size > 0 */
-static struct node *init_node(const void *key, unsigned long long key_size,
-                              const void *value,
-                              unsigned long long value_size) {
+static struct node *init_node(const void *key, size_t key_size, const void *value, size_t value_size) {
   struct node *node = calloc(1, sizeof *node);
   if (!node) return NULL;
 
@@ -122,8 +116,7 @@ static struct node *init_node(const void *key, unsigned long long key_size,
 /* used internally to replace an existing mapping for a certain key. returns a
  * pointer to the previous key which has to be free'd. the
  * function assumes the node passed in isn't NULL */
-static void *node_replace_value(struct node *node, const void *value,
-                                unsigned long long value_size) {
+static void *node_replace_value(struct node *node, const void *value, size_t value_size) {
   void *old_value = node->value;
   if (value_size) {
     unsigned char *tmp_value = calloc(value_size, 1);
@@ -140,9 +133,7 @@ static void *node_replace_value(struct node *node, const void *value,
 
 /* used internally to prepend a bucket to an entry. retuns true on success,
  * NULL on failure */
-static bool entry_prepend(struct entry *entry, const void *key,
-                          unsigned long long key_size, const void *value,
-                          unsigned long long value_size) {
+static bool entry_prepend(struct entry *entry, const void *key, size_t key_size, const void *value, size_t value_size) {
   struct node *node = init_node(key, key_size, value, value_size);
   if (!node) return false;
 
@@ -159,9 +150,9 @@ static bool entry_prepend(struct entry *entry, const void *key,
 /* used internally to check whether an entry contains a mapping for a certain
  * key. returns a pointer to the node which contains the same key, or NULL if
  * no such node found */
-static struct node *entry_contains(struct entry *entry, const void *key,
-                                   int (*cmpr)(const void *key,
-                                               const void *other)) {
+static struct node *entry_contains(struct entry *entry,
+                                   const void *key,
+                                   int (*cmpr)(const void *key, const void *other)) {
   if (!entry->head) return NULL;
   for (struct node *tmp = entry->head; tmp; tmp = tmp->next) {
     if (cmpr(key, tmp->key) == 0) return tmp;
@@ -175,14 +166,13 @@ static bool resize_table(struct hash_table *table) {
   if (!table) return false;
   if (!table->entries) return false;
 
-  unsigned long long new_capacity =
-      vector_resize(table->entries, table->capacity << TABLE_GROWTH);
+  size_t new_capacity = vector_resize(table->entries, table->capacity << TABLE_GROWTH);
   if (new_capacity == table->capacity) return false;
 
   table->capacity = new_capacity;
 
   // rehash every key-value pair
-  for (unsigned long long pos = 0; pos < table->capacity; pos++) {
+  for (size_t pos = 0; pos < table->capacity; pos++) {
     struct entry *entry = vector_at(table->entries, pos);
     if (!entry->head) continue;  // entry is empty
 
@@ -200,8 +190,7 @@ static bool resize_table(struct hash_table *table) {
         entry->head->prev = NULL;
       }
 
-      unsigned long long new_pos =
-          hash(tmp->key, tmp->key_size) % table->capacity;
+      size_t new_pos = hash(tmp->key, tmp->key_size) % table->capacity;
       struct entry *new_entry = vector_at(table->entries, new_pos);
 
       if (!new_entry->head) {  // new entry is empty
@@ -224,9 +213,7 @@ static bool resize_table(struct hash_table *table) {
   return true;
 }
 
-void *table_put(struct hash_table *table, const void *key,
-                unsigned long long key_size, const void *value,
-                unsigned long long value_size) {
+void *table_put(struct hash_table *table, const void *key, size_t key_size, const void *value, size_t value_size) {
   if (!table) return NULL;
   if (!table->entries) return NULL;
   if (!key && !key_size) return NULL;
@@ -237,15 +224,13 @@ void *table_put(struct hash_table *table, const void *key,
   }
 
   // get the entry index from the hash
-  unsigned long long pos = hash(key, key_size) % table->capacity;
+  size_t pos = hash(key, key_size) % table->capacity;
   struct entry *entry = vector_at(table->entries, pos);
   if (!entry) return NULL;
 
   // there's an existing mapping for this key
   struct node *contains_same_key = entry_contains(entry, key, table->cmpr);
-  if (contains_same_key) {
-    return node_replace_value(contains_same_key, value, value_size);
-  }
+  if (contains_same_key) { return node_replace_value(contains_same_key, value, value_size); }
 
   // there isn't an existing mapping for this key
   bool success = entry_prepend(entry, key, key_size, value, value_size);
@@ -255,13 +240,12 @@ void *table_put(struct hash_table *table, const void *key,
   return NULL;
 }
 
-void *table_remove(struct hash_table *table, const void *key,
-                   unsigned long long key_size) {
+void *table_remove(struct hash_table *table, const void *key, size_t key_size) {
   if (!table) return NULL;
   if (!table->entries) return NULL;
   if (!key && !key_size) return NULL;
 
-  unsigned long long pos = hash(key, key_size) % table->capacity;
+  size_t pos = hash(key, key_size) % table->capacity;
   struct entry *entry = vector_at(table->entries, pos);
   if (!entry) return NULL;
 
@@ -284,9 +268,7 @@ void *table_remove(struct hash_table *table, const void *key,
 
   unsigned char *old_value = removed->value;
 
-  if (table->destroy_key) {
-    table->destroy_key(removed->key);
-  }
+  if (table->destroy_key) { table->destroy_key(removed->key); }
   free(removed->key);
   free(removed);
   table->num_of_elements--;
@@ -294,13 +276,12 @@ void *table_remove(struct hash_table *table, const void *key,
   return old_value;
 }
 
-void *table_get(struct hash_table *table, const void *key,
-                unsigned long long key_size) {
+void *table_get(struct hash_table *table, const void *key, size_t key_size) {
   if (!table) return NULL;
   if (!table->entries) return NULL;
   if (!key && !key_size) return NULL;
 
-  unsigned long long pos = hash(key, key_size) % table->capacity;
+  size_t pos = hash(key, key_size) % table->capacity;
   struct entry *entry = vector_at(table->entries, pos);
   if (!entry) return NULL;
